Sync c-h nrpe.py file to fix nrpe files copy issue
Due to a change to PY3, the charmhelpers directory is not in the
CHARMDIR rather than in CHARMDIR/hooks. This broke the
copy_nrpe_checks() function in c-h. c-h has been updated, to look in
both CHARMDIR and CHARMDIR/hooks for the charmhelpers directory, and
this patchset includes the new function definition.

Note: the change in unit_tests/test_actions.py is to cope with a
difference between how subprocess.CalledProcessError() formats strings
when using str() between py35 and py36.  Python3.6 adds a period/dot to
the end of the sentence that Python3.5 didn't.

Change-Id: I845169c839bd2058e10076f24bf5393db0ec218a
Closes-Bug: #1796830